+++ CHANGELOG.md
@@ -4,6 +4,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add `stickyProps` to `HeaderNav`.
+- Feature: Add `useDeepCompareEffect` hook.
+- Feature: Add `useWindowWidth` hook.
+- Fix: Pass `isMenuExpanded` prop to styled component.
+- Fix: Fix `Dropdown` position if content is wider than the window.
 - Feature: New `setting` prop for `Text` component and new `lineHeight` options.
 
 ## [2.80.0] - 2020-07-13